reorganize fragment code
